Separated functions into specific files
